Add cli arg to remove plt normalization
By default imshow(img) normalizes img, which is can be undesirable when
trying to visualize the error, for example. In other times it can be
very helpful in order to determine *where* the model gets the most
wrong. The default behavior is unchanged, and to view the unnormalized
image simply pass the --no-norm arg